rename beautyCta->cta
