+++ assets/src/edit-story/stickers/beauty/cta.js
